new error infrastructure (#333)
* new error infrastructure

* fail compile test alterations for x86_64 profile

* fail test alterations for x86_64 profile

* error codes alterations for new error infrastructure

* update script to build opam dependencies

* fix missing path entry

* fix unit-fail-compilation tests by adding -Werror

* add csv renderer

* improve performance with nailgun and fix a few bugs

* make eclipse plugin work again

* fix typo in -Wno-unspecified and add tests

* add documentation for rv-error

+++ tests/unit-fail-compilation/j055f.c
@@ -1,5 +1,5 @@
 int x;
-enum e {a = (int)&x};
+enum e {a = (long)&x};
 int main(void){
 
 }
