+++ tests/unit-fail/j037e.c
@@ -1,10 +1,10 @@
 int main(void) {
-      int x = 10;
-      int* a;
-      long* b;
+      long x = 10;
+      long* a;
+      long long* b;
 
       a = &x;
-      b = (long*)a;
+      b = (long long*)a;
       *b;
       return 0;
 }
