+++ tests/unit-fail/j037h.c
@@ -1,9 +1,9 @@
 #include <stdlib.h>
 
 int main(void) {
-      int *p = malloc(sizeof(int));
+      long *p = malloc(sizeof(long));
 
-      *(long*)p = 42;
+      *(long long*)p = 42;
 
       return *p;
 }
