+++ tests/unit-fail/z00024.c
@@ -1,6 +1,7 @@
 // kcc-assert-error(00024a)
+#include<stdint.h>
 int main(void){
 	int x = 0;
-	return (int)(&x) & 7 + 132;
+	return (intptr_t)(&x) & 7 + 132;
 }
