+++ tests/unit-fail/z00024a.c
@@ -1,5 +1,6 @@
+#include<stdint.h>
 // kcc-assert-error(00024a)
 int main(void){
 	int x = 0;
-	return (int)(&x) & -7 + 146;
+	return (intptr_t)(&x) & -7 + 146;
 }
